+++ internal/repository/delete.go
@@ -1,6 +1,6 @@
 package repository
 
-func (r *SQLiteUserRepository) DeleteUserByID(id int) error {
+func (r *SQLiteUserRepository) DeleteUserByID(id float64) error {
 	_, err := r.db.Exec(`
 		DELETE FROM users
 		WHERE id = ?
